requirements.txt
fastapi
uvicorn
langchain
langgraph
pydantic
openai
python-dotenv
chromadb
tiktoken
langchain-openai
langchain-chroma
